fix: 나라 목록 API 요청 여러번 하는 버그 수정
