+++ src/pages/cahoots/create.tsx
@@ -1,23 +1,20 @@
-import { useRef, useState } from 'react';
+import { useState } from 'react';
 import { useForm, FieldErrors } from 'react-hook-form';
 
 import { useRouter } from 'next/router';
 
+import CountrySelectBox from '@/components/cahoot/CountrySelectBox';
 import DateRangeInput from '@/components/common/DateRangeInput';
 import FormItem from '@/components/common/FormItem';
 import ImageUpload from '@/components/common/ImageUpload';
 import Layout from '@/components/common/Layout';
 import NumberInput from '@/components/common/NumberInput';
 import RadioBtn from '@/components/common/RadioBtn';
-import SelectBox from '@/components/common/SelectBox';
 import TextArea from '@/components/common/TextArea';
 import TextInput from '@/components/common/TextInput';
 import HeaderWithBackButton from '@/components/mypage/HeaderWithBackButton';
 import PlaceSearchBar from '@/components/PlaceSearchBar';
-import { useSuspendedQuery } from '@/hooks/useSuspendedQuery';
 import wrapper, { useTypeSelector } from '@/store';
-import { CountriesType } from '@/types/cahoot';
-import { Response } from '@/types/response';
 // import Map from '@/components/Map';
 
 const positionOption = [
@@ -51,9 +48,9 @@ export interface FormDataType {
 }
 
 export default function CreateCahoot() {
+  const [selectedCountry, setSelectedCountry] = useState('국가'); // 휴양지 위치
   const token = useTypeSelector((state) => state.user.token);
   const router = useRouter();
-  const [selectedCountry, setSelectedCountry] = useState('국가'); // 휴양지 위치
   const {
     register,
     handleSubmit,
@@ -90,18 +87,6 @@ export default function CreateCahoot() {
 
   const onErrors = (errors: FieldErrors) => {
     console.log(errors);
-  };
-
-  const {
-    data: {
-      data: { result: countries },
-    },
-  } = useSuspendedQuery<Response<CountriesType>>(['MarketCountries'], () =>
-    fetch(`${process.env.NEXT_PUBLIC_HOST}/api/markets/countries`).then((res) => res.json())
-  );
-
-  const changeCountry = (country: string) => {
-    setSelectedCountry(country);
   };
 
   return (
@@ -211,16 +196,12 @@ export default function CreateCahoot() {
                   required: true,
                 })}
               />
-              <SelectBox
-                items={countries}
+              <CountrySelectBox
                 currentItem={selectedCountry}
-                changeItem={changeCountry}
-                size="large"
-                name="country"
+                selectItem={setSelectedCountry}
                 setValue={setValue}
                 trigger={trigger}
               />
-              {/* TODO: Loading 일 때 처리 */}
               <input
                 id="location"
                 type="hidden"
